Refactor `spotlight-hero` component to TypeScript, add responsive image support, and improve type definitions.
